+++ inst/editor/src/state/app_info.ts
@@ -2,8 +2,8 @@ import type { PayloadAction } from "@reduxjs/toolkit";
 import { createSlice } from "@reduxjs/toolkit";
 import type { MessageToClientByPath } from "communication-types";
 import type { App_Info } from "communication-types/src/AppInfo";
-import type { Raw_R_Info } from "r-ast-parsing";
-import { raw_R_info_to_app_info } from "r-ast-parsing/src/raw_R_info_to_app_info";
+import type { Raw_R_Info } from "r-bindings";
+import { raw_R_info_to_app_info } from "r-bindings";
 import { useSelector } from "react-redux";
 import type { ShinyUiNode } from "ui-node-definitions/src/ShinyUiNode";
 import type { PlaceNodeArguments } from "ui-node-definitions/src/TreeManipulation/placeNode";
