+++ inst/r-bindings/src/flatten_arrays_and_lists.ts
@@ -1,9 +1,10 @@
-import type { Leaf_Node, Primatives, R_AST, R_AST_Node } from "r-ast-parsing";
-import { is_ast_branch_node } from "r-ast-parsing/src/node_identity_checkers";
-import { Parsing_Error } from "r-ast-parsing/src/parsing_error_class";
 import type { UnknownUiNode } from "ui-node-definitions/src/internal/unknown_code";
 
+import type { Primatives, Leaf_Node, R_AST, R_AST_Node } from ".";
+
 import { create_unknownUiFunction } from "./create_unknownUiFunction";
+import { is_ast_branch_node } from "./node_identity_checkers";
+import { Parsing_Error } from "./parsing_error_class";
 
 type Primative_Map = Record<string, Primatives>;
