+++ inst/r-bindings/src/raw_R_info_to_app_info.ts
@@ -3,8 +3,9 @@ import type {
   Script_Generation_Template,
 } from "communication-types/src/AppInfo";
 import type { Script_Position } from "communication-types/src/MessageToBackend";
-import type { Raw_R_Info } from "r-ast-parsing";
 import { SCRIPT_LOC_KEYS } from "ui-node-definitions/src/code_generation/generate_ui_script";
+
+import type { Raw_R_Info } from ".";
 
 import { ast_to_ui_node } from "./ast_to_shiny_ui_node";
 import {
