+++ inst/r-bindings/src/text_nodes/textNodeParsing.test.ts
@@ -1,6 +1,6 @@
-import type { Branch_Node, R_AST_Node } from "r-ast-parsing";
 import type { ShinyUiParentNode } from "ui-node-definitions/src/ShinyUiNode";
 
+import type { R_AST_Node, Branch_Node } from "..";
 import { ast_to_ui_node } from "../ast_to_shiny_ui_node";
 
 import { is_text_node } from "./is_text_node";
